requirements.txt
Django==1.10.3
click==6.6
dj-database-url==0.4.1
django-cors-headers==1.3.1
djangorestframework==3.5.3
drf-nested-routers==0.11.1
factory-boy==2.7.0
fake-factory==0.7.2
psycopg2==2.6.2
python-dateutil==2.6.0
python-dotenv==0.6.1
six==1.10.0
